AI Voice Chatbot/requirements.txt
Flask
pyaudio
requests
speechrecognition
gTTS
pygame
langdetect
python-dotenv
google-generativeai
sarvamai